# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ordnung",
    "description": "Local API of Ordnung, the private AI secretary for life admin.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/drafts": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "List Drafts",
        "description": "Every letter, newest first.",
        "operationId": "list_drafts_api_drafts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Draft"
                  },
                  "type": "array",
                  "title": "Response List Drafts Api Drafts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Create Draft",
        "description": "Draft a cancellation, objection, reply or template letter (withdrawal, more time, instalments,\ndefect, data access, receipts, deposit, new address): fixed legal wording, model-written courtesy\ntext and translation, automatic checks and \"how to send it\".",
        "operationId": "create_draft_api_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "Get Draft",
        "description": "One letter.",
        "operationId": "get_draft_api_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "drafts"
        ],
        "summary": "Update Draft",
        "description": "Edit the letter; the checks (placeholders, references, dates …) run again.",
        "operationId": "update_draft_api_drafts__draft_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "drafts"
        ],
        "summary": "Delete Draft",
        "description": "Delete a letter, with its proofs and their files.",
        "operationId": "delete_draft_api_drafts__draft_id__delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/translate": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Translate Draft",
        "description": "Translate the letter again, as it stands after your edits (only the translation changes).",
        "operationId": "translate_draft_api_drafts__draft_id__translate_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "409": {
            "description": "The zero-token demo can't ask the model for a new translation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/pdf": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "Draft Pdf",
        "description": "The letter as a printable DIN 5008 PDF.",
        "operationId": "draft_pdf_api_drafts__draft_id__pdf_get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/sent": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Draft Sent",
        "description": "Record that the letter was sent (channel and day) and add a follow-up to-do.",
        "operationId": "draft_sent_api_drafts__draft_id__sent_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkSentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Draft": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "cancellation",
              "objection",
              "general_reply",
              "withdrawal",
              "extension_request",
              "payment_plan",
              "defect_notice",
              "data_access",
              "receipts_inspection",
              "deposit_return",
              "address_change"
            ],
            "title": "Kind"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "de"
          },
          "party_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "contract_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Id"
          },
          "sender_block": {
            "type": "string",
            "title": "Sender Block",
            "default": ""
          },
          "recipient_block": {
            "type": "string",
            "title": "Recipient Block",
            "default": ""
          },
          "place_date": {
            "type": "string",
            "title": "Place Date",
            "default": ""
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "default": ""
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "body_translation": {
            "type": "string",
            "title": "Body Translation",
            "default": ""
          },
          "enclosures": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enclosures"
          },
          "notes_for_user": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes For User"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/DraftCheck"
            },
            "type": "array",
            "title": "Checks"
          },
          "send_guidance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SendGuidance"
              },
              {
                "type": "null"
              }
            ]
          },
          "sent_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent Channel"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "final",
              "sent"
            ],
            "title": "Status",
            "default": "draft"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "tracking_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "language",
          "party_id",
          "case_id",
          "doc_id",
          "contract_id",
          "sender_block",
          "recipient_block",
          "place_date",
          "subject",
          "body",
          "body_translation",
          "enclosures",
          "notes_for_user",
          "checks",
          "send_guidance",
          "sent_channel",
          "status",
          "sent_at",
          "tracking_number",
          "created_at",
          "updated_at"
        ],
        "title": "Draft"
      },
      "DraftCheck": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "ok",
          "detail"
        ],
        "title": "DraftCheck"
      },
      "DraftCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "cancellation",
              "objection",
              "general_reply",
              "withdrawal",
              "extension_request",
              "payment_plan",
              "defect_notice",
              "data_access",
              "receipts_inspection",
              "deposit_return",
              "address_change"
            ],
            "title": "Kind"
          },
          "party_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Id"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "contract_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id",
            "description": "accepted for symmetry; the thread follows the letter"
          },
          "instructions": {
            "type": "string",
            "maxLength": 4000,
            "title": "Instructions",
            "default": ""
          },
          "language": {
            "type": "string",
            "enum": [
              "de",
              "en"
            ],
            "title": "Language",
            "default": "de"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LetterDetails"
              },
              {
                "type": "null"
              }
            ],
            "description": "the facts a template letter needs (withdrawal, payment plan …)"
          },
          "suspend_enforcement": {
            "type": "boolean",
            "title": "Suspend Enforcement",
            "description": "an objection also applies to suspend enforcement (einstweilige Einstellung at a court, Aussetzung der Vollziehung at an authority); ignored for other letters and a court payment order",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "DraftCreate",
        "description": "What to write: the kind of letter and what it is about."
      },
      "DraftPatch": {
        "properties": {
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "body_translation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Translation"
          },
          "sender_block": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Block"
          },
          "recipient_block": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Block"
          },
          "place_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Date"
          },
          "enclosures": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enclosures"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "draft",
                  "final"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DraftPatch",
        "description": "Edits the person makes to a draft (the checks run again)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LetterDetails": {
        "properties": {
          "subject_matter": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Matter",
            "description": "what was ordered or agreed, e.g. 'Kaffeemaschine'"
          },
          "ordered_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordered On",
            "description": "the day the contract was concluded"
          },
          "received_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received On",
            "description": "the day the goods arrived"
          },
          "instructions_missing": {
            "type": "boolean",
            "title": "Instructions Missing",
            "description": "no (or wrong) instructions about the right of withdrawal were given",
            "default": false
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline",
            "description": "the deadline that should be extended"
          },
          "until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until",
            "description": "the new date asked for"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount",
            "description": "the total owed, or the deposit"
          },
          "instalment": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Instalment",
            "description": "the monthly instalment offered"
          },
          "first_instalment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Instalment",
            "description": "the day of the first instalment"
          },
          "defect": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Defect",
            "description": "what is broken or wrong"
          },
          "noticed_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Noticed On",
            "description": "since when the defect exists"
          },
          "fix_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fix By",
            "description": "the day by which it should be repaired"
          },
          "period": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Period",
            "description": "the billing period"
          },
          "moved_out_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moved Out On",
            "description": "the day the flat was handed back"
          },
          "moved_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moved On",
            "description": "the day of the move"
          },
          "old_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Address"
          },
          "new_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "New Address"
          },
          "recipient": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient",
            "description": "name and address of a recipient not in Ordnung yet"
          }
        },
        "type": "object",
        "title": "LetterDetails",
        "description": "Facts a template letter needs besides the letter, contract or person it is about.\n\nEverything is optional here; each template names the facts it requires\n(:data:`ordnung.drafts.templates.TEMPLATES`). Dates are ISO ``YYYY-MM-DD`` (anything else is refused\nwith a clear message, never a server error), amounts in euros."
      },
...
      "MarkSentRequest": {
        "properties": {
          "channel": {
            "type": "string",
            "minLength": 1,
            "title": "Channel"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "tracking_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number",
            "description": "the Einschreiben's number (its check digit is checked)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "channel",
          "date"
        ],
        "title": "MarkSentRequest",
        "description": "How and when the letter was sent."
      },
...
      "SendGuidance": {
        "properties": {
          "send_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Send By"
          },
          "must_arrive_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Must Arrive By"
          },
          "form": {
            "type": "string",
            "enum": [
              "text_form",
              "written_form",
              "any"
            ],
            "title": "Form",
            "default": "text_form"
          },
          "form_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Note"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/SendChannel"
            },
            "type": "array",
            "title": "Channels"
          },
          "tips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tips"
          }
        },
        "type": "object",
        "required": [
          "send_by",
          "must_arrive_by",
          "form",
          "form_note",
          "channels",
          "tips"
        ],
        "title": "SendGuidance"
      },
...
    }
  }
}